+++ swing/common-model/src/main/java/is/codion/swing/common/model/component/combobox/FilterComboBoxModel.java
@@ -59,16 +59,18 @@ public interface FilterComboBoxModel<T> extends FilterModel<T>, ComboBoxModel<T>
 	 * If this model already contains the item, calling this method has no effect.
 	 * Note that if the item does not fulfill the visible predicate, it will be filtered right away.
 	 * @param item the item to add
+	 * @return true if the item is visible
 	 * @throws IllegalArgumentException in case the item fails validation
 	 * @see Items.Visible#predicate()
 	 */
-	void addItem(T item);
+	boolean addItem(T item);
 
 	/**
 	 * Removes the given item from this model
 	 * @param item the item to remove
+	 * @return true if the item was visible
 	 */
-	void removeItem(T item);
+	boolean removeItem(T item);
 
 	/**
 	 * Replaces the given item in this combo box model